host/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "host"
version = "0.0.1"
requires-python = ">=3.9"
dependencies = [
  "core",
  "enclave",
  "cbor2==5.2.0",
  "cose==0.9.dev2",
  "pynacl==1.5.0",
  "pycryptodome==3.19.1",
  "pyOpenSSL==24.0.0",
  "base58==2.1.1",
]

[project.scripts]
verify_attestation = "host.verify_attestation:main"
verify_query = "host.verify_query:main"
forward = "host.forward:main"
https_forward = "host.https_forward:main"